+++ discord/commands/__init__.py
@@ -25,3 +25,4 @@ DEALINGS IN THE SOFTWARE.
 
 from .commands import *
 from .context import *
+from .errors import *
