+++ discord/commands/commands.py
@@ -27,7 +27,6 @@ from __future__ import annotations
 
 import asyncio
 import types
-from discord.types.channel import TextChannel, VoiceChannel
 import functools
 import inspect
 from collections import OrderedDict
@@ -39,8 +38,9 @@ from ..user import User
 from ..message import Message
 from .context import ApplicationContext
 from ..utils import find, get_or_fetch, async_all
-from ..errors import DiscordException, NotFound, ValidationError, ClientException
+from ..errors import ValidationError, ClientException
 from .errors import ApplicationCommandError, CheckFailure, ApplicationCommandInvokeError
+from .permissions import Permission, has_role, has_any_role, is_user, is_owner, permission
 
 __all__ = (
     "_BaseCommand",
@@ -959,99 +959,3 @@ def validate_chat_input_description(description: Any):
         raise ValidationError(
             "Description of a chat input command must be less than 100 characters and non empty."
         )
-
-# Slash Command Permissions
-class Permission:
-    def __init__(self, id: Union[int, str], type: int, permission: bool = True, guild_id: int = None):
-        self.id = id
-        self.type = type
-        self.permission = permission
-        self.guild_id = guild_id
-
-    def to_dict(self) -> Dict[int, int, bool]:
-        return {"id": self.id, "type": self.type, "permission": self.permission}
-
-def permission(role_id: int = None, user_id: int = None, permission: bool = True, guild_id: int = None):
-    def decorator(func: Callable):
-        if not role_id is None:
-            app_cmd_perm = Permission(role_id, 1, permission, guild_id)
-        elif not user_id is None:
-            app_cmd_perm = Permission(user_id, 2, permission, guild_id)
-        else:
-            raise ValueError("role_id or user_id must be specified!")
-
-        # Create __app_cmd_perms__
-        if not hasattr(func, '__app_cmd_perms__'):
-            func.__app_cmd_perms__ = []
-
-        # Append
-        func.__app_cmd_perms__.append(app_cmd_perm)
-
-        return func
-
-    return decorator
-
-def has_role(item: Union[int, str], guild_id: int = None):
-    def decorator(func: Callable):
-        # Create __app_cmd_perms__
-        if not hasattr(func, '__app_cmd_perms__'):
-            func.__app_cmd_perms__ = []
-
-        # Permissions (Will Convert ID later in register_commands if needed)
-        app_cmd_perm = Permission(item, 1, True, guild_id) #{"id": item, "type": 1, "permission": True}
-
-        # Append
-        func.__app_cmd_perms__.append(app_cmd_perm)
-
-        return func
-
-    return decorator
-
-def has_any_role(*items: Union[int, str], guild_id: int = None):
-    def decorator(func: Callable):
-        # Create __app_cmd_perms__
-        if not hasattr(func, '__app_cmd_perms__'):
-            func.__app_cmd_perms__ = []
-
-        # Permissions (Will Convert ID later in register_commands if needed)
-        for item in items:
-            app_cmd_perm = Permission(item, 1, True, guild_id) #{"id": item, "type": 1, "permission": True}
-
-            # Append
-            func.__app_cmd_perms__.append(app_cmd_perm)
-
-        return func
-
-    return decorator
-
-def is_user(user: int, guild_id: int = None):
-    def decorator(func: Callable):
-        # Create __app_cmd_perms__
-        if not hasattr(func, '__app_cmd_perms__'):
-            func.__app_cmd_perms__ = []
-
-        # Permissions (Will Convert ID later in register_commands if needed)
-        app_cmd_perm = Permission(user, 2, True, guild_id) #{"id": user, "type": 2, "permission": True}
-
-        # Append
-        func.__app_cmd_perms__.append(app_cmd_perm)
-
-        return func
-
-    return decorator
-
-def is_owner(guild_id: int = None):
-    def decorator(func: Callable):
-        # Create __app_cmd_perms__
-        if not hasattr(func, '__app_cmd_perms__'):
-            func.__app_cmd_perms__ = []
-
-        # Permissions (Will Convert ID later in register_commands if needed)
-        app_cmd_perm = Permission("owner", 2, True, guild_id) #{"id": "owner", "type": 2, "permission": True}
-
-        # Append
-        func.__app_cmd_perms__.append(app_cmd_perm)
-
-        return func
-
-    return decorator
